% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFC_avl.R
\name{IFC_avl}
\alias{IFC_avl}
\title{IFC Package Checker}
\usage{
IFC_avl()
}
\value{
a invisible logical
}
\description{
Test if IFC package is installed.
}
\examples{
IFC_installed <- IFC_avl()
print(IFC_installed)
}
