% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFCdata-package.R
\docType{data}
\name{IFCdata-package}
\alias{IFCdata-package}
\alias{IFCdata}
\title{Data Package for Imaging Flow Cytometry}
\description{
Contains Imaging Flow Cytometry data.
}
\author{
Maintainer: Yohann Demont <git.demont@gmail.com>
}
